更新图形系统中UICanva画布API功能 相关的Issue
原因（目的、解决的问题等）
更新图形系统中UICanva画布API功能
RM008 对标H5，支持实现如下属性及接口
RM009 线条样式属性及矩形方法
RM011 在画布上绘制文本（无填充）以及向画布上绘制图像
RM012 支持alpha和合成模式属性
RM013 渐变填充方法的实现

描述（做了什么，变更了什么）
更新图形系统中UICanva画布API功能
RM008 对标H5，支持实现如下属性及接口
RM009 线条样式属性及矩形方法
RM011 在画布上绘制文本（无填充）以及向画布上绘制图像
RM012 支持alpha和合成模式属性
RM013 渐变填充方法的实现

测试用例（新增、改动、可能影响的功能）
foundation/graphic/ui/test/unittest/components/ui_canvas_unit_test.cpp

+++ test/uitest/test_qrcode/ui_test_qrcode.cpp
@@ -155,7 +155,7 @@ void UITestQrcode::UIKitUIQrcodeTestDisplay005()
         qrcode->SetPosition(20, 30, 250, 250); // 20 x 30 y 250 width 250 height
         const int16_t length = 2050; // 2050 length
         char text[length];
-        for (int i = 0; i < length - 1; i++) {
+        for (int16_t i = 0; i < length - 1; i++) {
             text[i] = 'd';
         }
         text[length - 1] = '\0';
